requirements/requiremenrsWorker.txt
arq
asyncpg
sqlalchemy
pandas
pydantic
fastapi
fastapi-users[sqlalchemy]
PyPDF2
python-docx
openpyxl 
transformers
aioboto3